better error for incompatible option types
previously, if your option had a type (not an instance) as the default
value the error message claimed to want to convert to "<class 'type'>",
which was of course both untrue and unhelpful.

example:

options = {
	'foo': dict,
}

urd.build('foo', foo='bar')

accelerator.deptree.OptionException: Failed to convert option foo of
<class 'str'> to <class 'dict'> on method foo

these messages (and the code!) could do with more improvement, e.g. if
you have values of a bad type inside a dict the message doesn't make
that clear.