[shell] expand ${VAR} in alias expansion
so you can have e.g.

[alias]
	grep = grep ${AXGREP}

and if AXGREP="--ordered --headers" those two arguments will be set
when calling "ax grep". (So shell splitting is applied to the variable,
and unset variables expand to no words.)

Only the ${NAME} syntax is allowed, and only as a stand alone word.
Because of limitations in the shlex module, there is no support for
escaping these, which is why these restrictions on where vars expand
were choosen.