[status] preserve status stack when missing exception
the problem this avoids is if you have something like this:

	for blah in ds.iterate(...):
		1/0

you would expect to see "Iterating ds" in the status stack shown with
the exception, but because the exception is happening after the function
that set the status the context manager doesn't see the exception.

finally clauses however always run.