Dataset.merge() worries less about hashlabel
just using it from the other ds if set, and otherwise keeping it unless
overwritten.

This obviously gives you some extra rope to hang yourself, but it turns out
I need this functionality, and if they are actually hashed differently you
will probably be saved by the line counts not matching.