[dataset_type] + datetime '%?' to make one element optional
or several with a COUNT/RANGE.

RANGE specifies minimum and maximum matching tokens, COUNT is the same as
setting LOW and HIGH the same.

If less than LOW token match, no part of the string is consumed, no updates
occur to the parsed value and matching continues after HIGH elements in the
format string.