[csvimport] comments don't affect slicing
so now the first non-skipped line always goes to slice 0 and so on.
(and the same rule applies to skipped lines.)

bad lines however are not skipped, they go in the slice a good line at
the same position would have gone (and the following line goes in the
next slice).

but as long as you don't enable allow_bad the lines are now distributed
"perfectly", and dataset_unroundrobin can successfully recreate the
original order.