[ax grep] more preparation for --unique.
self.write() is so the written data can be preprocessed (to remove
duplicates) at the time of writing, as doing it earlier would make it
non-deterministic which lines were chosen.

The merge_buffer stuff is so a more complex data structure can be used.
(To keep everything around until .write().)

Adding .put_local to LocalQueue is so even more special casing is not
needed, and --unique can always be treated as --ordered, even with only a
single slice.

Again, this commit changes no functionality.