[job] add support for pathlib.Path for modern python
The whole Job API and its internal structure assume filename objects
to be strings. However, most methods either do implicit conversions
to strings, or use os.path methods, that since Python3.6 support
pathlib.Path objects. In consequence, it is only necessary to
explicitly convert inputs to strings when there are internal checks
that assume the object to be a string