accelerator conf (not project conf) can [include]
because of configparser limitations they can not be included at the point
where they are specified, so everything in the parent file overrides the
same option in the included file.

several files are supported by specifying several (shell) words in the
value, as configparser does not support repeating keys (without discarding
values).