[ax grep] ^T works with --roundrobin.
For some value of "works".

The callback that updates the status (grep -> mk_iter -> cb) is called
separately by each iterated slice, which means that you get
len(want_slices) updates at the same time as long as all slices are going.

This doesn't break anything, it just makes updates slow (in balanced
datasets).

When double pressing ^T it still prints as if there was a slice per
process, which may confuse.