[dsutil] trivial optimisation on 32bit archs
when reading small numbers.

This doesn't matter, since no one who cares about performance will be
using a 32bit arch today, but I spotted it and just couldn't resist.

On 64bit archs pyInt_FromS32 and pyInt_FromS64 are both PyInt_FromLong, so
this has no effect on the compiled code, but on 32bit pyInt_FromS64 is
PyLong_FromLongLong which is presumably slower.