[urd] record (first) build_job that produces a list.
A new add that changes nothing except the build_job will compare equal to
the old one and therefore will not write anything to the log.

This does not introduce a new log line version, it just adds an extra
field at the end of a v4 line, that earlier versions will ignore.