[dsutil] try to force compiler to actually do float truncation
where needed for correct hashing.

The goal is to make WriteFloat32.hash() and WriteComplex32.hash() return
the same value as you would get first writing to that format and then
hashing.

My reading of the GCC documentation is that this should be the default
with -std=c99, but it clearly isn't. It also sounds like -ffloat-store
should force this without volatile, but it does not.