[dataset] allow (date)times with UTC, preserving it.
No other timezones are allowed. The definition of if the timezone is UTC
is that .utcoffset() is 0. I think that's the best we can do.

Will return objects with tzinfo=datetime.timezone.utc or our own class
depending on python version.

Older accelerator versions will ignore the flag when reading, unless
very old (before 2020-10-19).

The timezone is ignored in minmax calculations, so .min and .max will
always be naive objects.