[scripts] Skip tests on python 3.13.0.
There are bugs in the threading module causing the tests to fail. This
version is unfortunately in the last musl 1.1 containers, so we're stuck
with it until we ignore musl 1.1 completely.

(I've been removing it from the container before running the build for a
while, but after deciding that the problem is 3.13.0 and not something in
exax or 3.13 in general I'm going to skip the tests and hope for the best.)

The breakage is in the output test, as the threading module produces
strange errors and those tracebacks end up in the output. I'm not
going to change the test to accomodate that.

Here's a minimal example that triggers a threading bug. It's not the same
bug as the output tests dies from, but it's short.

import os, threading, time
if os.fork():
	time.sleep(0.1)

On a working python this will print nothing. On 3.13.0, at least as present
in the musllinux_1_1 containers, you get an AssertionError in
Thread._after_fork(). (No error happens if you skip the threading import.)